Moves types into their own files
