+++ model/CreateTablesResponse.go
@@ -1,0 +1,6 @@
+package model
+
+type CreateTablesResponse struct {
+	UnrecognizedProperties map[string][]string `json:"unrecognizedProperties"`
+	Status                 string              `json:"status"`
+}
