+++ model/GetTenantsResponse.go
@@ -1,0 +1,6 @@
+package model
+
+type GetTenantsResponse struct {
+	ServerTenants []string `json:"SERVER_TENANTS"`
+	BrokerTenants []string `json:"BROKER_TENANTS"`
+}
