+++ model/ValidateSchemaReponse.go
@@ -1,0 +1,6 @@
+package model
+
+type ValidateSchemaResponse struct {
+	Ok    bool
+	Error string
+}
